+++ shipPet/src/main/java/com/ship/service/impl/MsgServiceImpl.java
@@ -22,32 +22,32 @@ import java.util.List;
 public class MsgServiceImpl extends ServiceImpl<MsgMapper, Msg> implements IMsgService {
 
     @Override
-    public Long getByClient(Long clientId) {
+    public Integer getByClient(Integer clientId) {
         return baseMapper.getByClient(clientId);
     }
 
     @Override
-    public List<MsgVo> getADoctor(Long clientId, Long doctorId) {
+    public List<MsgVo> getADoctor(Integer clientId, Integer doctorId) {
         return baseMapper.getADoctor(clientId, doctorId);
     }
 
     @Override
-    public List<MsgNameVo> getDoctorGroup(Long clientId, Long doctorId) {
+    public List<MsgNameVo> getDoctorGroup(Integer clientId, Integer doctorId) {
         return baseMapper.getDoctorGroup(clientId, doctorId);
     }
 
     @Override
-    public Long getByDoctor(Long doctorId) {
+    public Integer getByDoctor(Integer doctorId) {
         return baseMapper.getByDoctor(doctorId);
     }
 
     @Override
-    public List<MsgVo> getAClient(Long doctorId, Long clientId) {
+    public List<MsgVo> getAClient(Integer doctorId, Integer clientId) {
         return baseMapper.getAClient(doctorId, clientId);
     }
 
     @Override
-    public List<MsgNameVo> getClientGroup(Long doctorId, Long clientId) {
+    public List<MsgNameVo> getClientGroup(Integer doctorId, Integer clientId) {
         return baseMapper.getClientGroup(doctorId, clientId);
     }
 }
