+++ shipPet/src/main/java/com/ship/vo/NameVo.java
@@ -3,5 +3,5 @@ package com.ship.vo;
 /**
  * 部分下拉框查询名字与对应id
  */
-public record NameVo(Long roleId, String roleName) {
+public record NameVo(Integer roleId, String roleName) {
 }
